# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gin swagger",
        "title": "Gin swagger",
        "contact": {
            "name": "nathan.lu",
            "url": "https://tedmax100.github.io/"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/etherscan/gettraderecord": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EtherScan"
                ],
                "summary": "Get Wallet Found",
                "parameters": [
                    {
                        "type": "string",
                        "description": "wallet",
                        "name": "wallet",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/etherscan/getwalletfound": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EtherScan"
                ],
                "summary": "Get Wallet Found",
                "parameters": [
                    {
                        "type": "string",
                        "description": "wallet",
                        "name": "wallet",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}